refactor: moved user exceptions from its package to unify the exception package
