+++ src/main/kotlin/data/repositories/user/UserRepositoryImpl.kt
@@ -2,7 +2,7 @@ package org.baghdad.data.repositories.user
 
 import org.baghdad.data.local.UserDataSource
 import org.baghdad.logic.model.entities.UserEntity
-import org.baghdad.logic.model.exceptions.user.UserNotFoundException
+import org.baghdad.logic.model.exceptions.UserNotFoundException
 import org.baghdad.logic.repositories.UserRepository
 import java.util.UUID
