+++ src/main/kotlin/logic/usecase/user/GetUserByUsernameUseCase.kt
@@ -1,7 +1,7 @@
 package org.baghdad.logic.usecase.user
 
 import org.baghdad.logic.model.entities.UserEntity
-import org.baghdad.logic.model.exceptions.user.InvalidUsernameException
+import org.baghdad.logic.model.exceptions.InvalidUsernameException
 import org.baghdad.logic.repositories.UserRepository
 
 class GetUserByUsernameUseCase(
