+++ src/main/kotlin/presentation/user/GetUserByUsernameUI.kt
@@ -1,7 +1,7 @@
 package org.baghdad.presentation.user
 
-import org.baghdad.logic.model.exceptions.user.InvalidUsernameException
-import org.baghdad.logic.model.exceptions.user.UserNotFoundException
+import org.baghdad.logic.model.exceptions.InvalidUsernameException
+import org.baghdad.logic.model.exceptions.UserNotFoundException
 import org.baghdad.logic.usecase.user.GetUserByUsernameUseCase
 import org.baghdad.presentation.input.Reader
 import org.baghdad.presentation.output.Viewer
