+++ src/test/kotlin/logic/usecase/user/GetUserByUsernameUseCaseTest.kt
@@ -5,8 +5,8 @@ import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import org.baghdad.logic.model.entities.UserEntity
 import org.baghdad.logic.model.enums.UserType
-import org.baghdad.logic.model.exceptions.user.InvalidUsernameException
-import org.baghdad.logic.model.exceptions.user.UserNotFoundException
+import org.baghdad.logic.model.exceptions.InvalidUsernameException
+import org.baghdad.logic.model.exceptions.UserNotFoundException
 import org.baghdad.logic.repositories.UserRepository
 import org.baghdad.logic.usecase.user.GetUserByUsernameUseCase
 import org.junit.jupiter.api.BeforeEach
