+++ src/test/kotlin/logic/usecase/user/UserValidatorUseCaseTest.kt
@@ -5,11 +5,11 @@ import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import org.baghdad.logic.model.enums.UserType
-import org.baghdad.logic.model.exceptions.user.InvalidNameException
-import org.baghdad.logic.model.exceptions.user.InvalidPasswordException
-import org.baghdad.logic.model.exceptions.user.InvalidUsernameException
-import org.baghdad.logic.model.exceptions.user.UnauthorizedException
-import org.baghdad.logic.model.exceptions.user.UserAlreadyExistsException
+import org.baghdad.logic.model.exceptions.InvalidNameException
+import org.baghdad.logic.model.exceptions.InvalidPasswordException
+import org.baghdad.logic.model.exceptions.InvalidUsernameException
+import org.baghdad.logic.model.exceptions.UnauthorizedException
+import org.baghdad.logic.model.exceptions.UserAlreadyExistsException
 import org.baghdad.logic.repositories.UserRepository
 import org.baghdad.logic.usecase.user.UserValidatorUseCase
 import org.junit.jupiter.api.BeforeEach
